Add hydraulic manipulator with base joint using a custom plastic sheet
of .0095 thickness in conjunction with custom .02 thickness metal plates.

Fix #4.

Fix #1 as now every joint is made of plastic <-> metal contact besides
the joints 0, 3 which have a clearance of .0006 inbetween the two
metal cylinders of the DIY heavy duty bearing, hardly enough to insert a
layer of plastic unless a kind of thin tape?

Fix #5: Add link 4, joint 4 and the tool coupling sheet.

Add some hydraulic cylinders (not attached).