Mechanics: Fix #11: Add 2m segment hydraulic manipulator. Fix #4.
Fix #1 as now every joint is made of plastic <-> metal contact besides
the joints 0, 3 which have a clearance of .0006 inbetween the two
metal cylinders of the DIY heavy duty bearing, hardly enough to insert a
layer of plastic unless a kind of thin tape or pouring fluid and let
it harden.

Enlarge links'|segments' lengths to 2m.

Fix #5: Specify link4, joint 4 and the tool coupling sheet.

Add some hydraulic cylinders (not attached).