Swap static, moving cylinder joint 0. Rm hydraulic pump platform.
Unfortunately joint 0,1 now collide (#13).

Rm hydraulic pump platform as it is moved inside the hobbit hole to
reduce noise and thus no longer needed (it's usefulness was questionable
anyway because the pump can also be located higher at the joint0 sheet
if the size was increased and it'd remove space and possible depth reach
as it meant joint 1 had to be placed higher - both on walker and fruit
nut vegetable gardener).